+++ src/Rdd.Web/Querying/IQueryParser.cs
@@ -1,13 +1,12 @@
 ﻿using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.Primitives;
-using Rdd.Domain.Helpers;
+using Microsoft.AspNetCore.Mvc.Abstractions;
 using Rdd.Domain.Models.Querying;
-using System.Collections.Generic;
 
 namespace Rdd.Web.Querying
 {
     public interface IQueryParser<TEntity> where TEntity : class
     {
         Query<TEntity> Parse(HttpRequest request, bool isCollectionCall);
+        Query<TEntity> Parse(HttpRequest request, ActionDescriptor action, bool isCollectionCall);
     }
 }